feat(frontend): adiciona componente de cadastro de proposta de estagio
