+++ lpweb20201-estagio-tcc-front/src/app/app-routing.module.ts
@@ -12,19 +12,21 @@ import { CadastrarPerfilComponent } from './perfil/cadastrar-perfil/cadastrar-pe
 import { EditarPerfilComponent } from './perfil/editar-perfil/editar-perfil.component';
 import { AbrirPropostaDeTCCComponent } from './proposta-de-tcc/abrir-proposta-de-tcc/abrir-proposta-de-tcc.component';
 import { CadastrarPropostaDeTccComponent } from './proposta-de-tcc/cadastrar-proposta-de-tcc/cadastrar-proposta-de-tcc.component';
+import { CadastrarPropostaDeEstagioComponent } from './proposta-estagio/cadastrar-proposta-de-estagio/cadastrar-proposta-de-estagio.component';
 
 
 const routes: Routes = [
   { path: 'login', component: LoginComponent },
-  { path: 'propostas-de-estagio', component: PropostaEstagioComponent},
   {
     path: 'inicio', component: InicioComponent, children: [
+      { path: 'propostas-de-estagio', component: PropostaEstagioComponent},
+      { path: 'proposta-de-estagio/cadastro', component: CadastrarPropostaDeEstagioComponent },
+
       { path: 'propostas-de-tcc', component: PropostaDeTCCComponent },
       { path: 'propostas-de-tcc/:id', component: AbrirPropostaDeTCCComponent },
       { path: 'proposta-de-tcc/cadastro', component: CadastrarPropostaDeTccComponent },
       { path: 'perfil', component: PerfilComponent},
       { path: 'sobre', component: SobreComponent },
-      
       { path: '', component: HomeInicioComponent }
     ]
   },
