+++ lpweb20201-estagio-tcc-front/src/app/inicio/inicio.component.html
@@ -9,7 +9,9 @@
             <div class="navbar-nav">
                 <a class="nav-item nav-link" [routerLink]="['/inicio']" routerLinkActive="active"
                     [routerLinkActiveOptions]="{exact:true}">Início</a>
-                <a class="nav-item nav-link" [routerLink]="['/inicio', 'propostas-de-tcc']" routerLinkActive="active"
+                <a class="nav-item nav-link" [routerLink]="['/inicio', 'propostas-de-estagio']" routerLinkActive="active"
+                    [routerLinkActiveOptions]="{exact:true}">Propostas de Estágio</a>
+                    <a class="nav-item nav-link" [routerLink]="['/inicio', 'propostas-de-tcc']" routerLinkActive="active"
                     [routerLinkActiveOptions]="{exact:true}">Propostas de TCC</a>
                 <a class="nav-item nav-link" [routerLink]="['/inicio', 'perfil']" routerLinkActive="active"
                     [routerLinkActiveOptions]="{exact:true}">Perfil</a>
