+++ lpweb20201-estagio-tcc-front/src/app/proposta-estagio/cadastrar-proposta-de-estagio/cadastrar-proposta-de-estagio.component.css
@@ -1,0 +1,3 @@
+.required{
+    color:red;
+}
