+++ lpweb20201-estagio-tcc-front/src/app/proposta-estagio/proposta-estagio.component.ts
@@ -1,10 +1,4 @@
 import { Component, OnInit } from '@angular/core';
-import { AuthService } from '../auth.service';
-import { ActivatedRoute } from '@angular/router';
-import { PropostaEstagioService } from '../proposta-estagio.service';
-import { delay } from 'rxjs/operators';
-import { OrientacaoService } from '../orientacao.service';
-
 
 @Component({
   selector: 'app-proposta-estagio',
@@ -13,53 +7,9 @@ import { OrientacaoService } from '../orientacao.service';
 })
 export class PropostaEstagioComponent implements OnInit {
 
-  constructor(
-    public proposta$: PropostaEstagioService,
-    private orientacao$: OrientacaoService,
-    public auth$: AuthService,
-    private router: ActivatedRoute,
-    ) {
+  constructor() {
   }
 
-  user:any;
-  orientacao = null;
-  titulo = null;
-  conceitos = null;
-  resultados = null;
-  minhasOrientacoes = null;
-  orientacoes: any;
-
-  error = null;
-  sucesso = null;
-
- ngOnInit(): void {
-  this.user = this.auth$.user();
-
-  if(this.user){
-    this.orientacao$.get().subscribe(dados => {this.orientacoes = dados.results;this.encontrarOrientacao();});
-  }
-
-  }
-  enviarDados(){
-    let dados = {
-      orientacao_id: this.orientacao,
-      titulo: this.titulo,
-      conceitos: this.conceitos,
-      resultados_esperados: this.resultados,
-      membros_da_banca: []
-    }
-    this.proposta$.cadastrar(dados).subscribe(() => {this.sucesso = true}, err => this.error = err.error);
-    console.log(dados)
-  }
-
-  encontrarOrientacao(){
-
-
-    this.minhasOrientacoes = this.orientacoes.filter((alunoO)=> alunoO.tipo === "estagio" && alunoO.aluno.id === this.user.id)
-
-
-    console.log(this.minhasOrientacoes)
-  }
+ ngOnInit(): void {}
 
 }
-
